Fix thin packfile before uploading it
This fixes previous buggy behavior where the generated pack index file does not
correspond to the uploaded pack file. When git-index-pack fixes a thin pack,
it outputs the fixed pack. So we have to use that instead of the original
(possibly thin) pack. This means that some objects will be included in the pack
that were deltified when they were pushed from the client; however
git-index-pack will not index a thin pack before fixing it, so we have to deal
with this at least until we have a packfile indexer that indexes a thin pack.